enum 분리
